# [Performance] [PostRector] Avoid unnecessary traverse() to count total namespaces on NameImportingPostRector