`Namespace_` stmt is always on first level stmt, so no need to use `BetterNodeFinder` to locate into deep node, just loop the first level stmts to improve performance.