@samsonasik 2024-07-26
Got 2 seconds faster for rector-src code base:

**Before**

```
➜  rector-src git:(main) time bin/rector process --clear-cache
 2251/2251 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   

bin/rector process --clear-cache  235.76s user 10.75s system 717% cpu 34.337 total
```

**After**

```
➜  rector-src git:(avoid-unnecessary-traverse) time bin/rector process --clear-cache
 2251/2251 [▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓] 100%

                                                                                                                        
 [OK] Rector is done!                                                                                                   
                                                                                                                        
bin/rector process --clear-cache  232.87s user 9.79s system 740% cpu 32.787 total
```